refactor(runtime): split engine_scheduler.cpp (2230 LOC) into prefill + decode-pipeline TUs; RAII for ragged metadata
engine_scheduler.cpp had doubled past its allowlist rationale (1074 -> 2230
code LOC) and carried three domains. Pure moves, byte-identical bodies
(verified by reconstruction diff: 6 whitespace lines at the cut edges):

- engine_prefill.cpp: step_prefill driver, step_prefill_one,
  prefill_allocate_kv_blocks_, prefill_upload_metadata_ (886 lines)
- engine_decode_pipeline.cpp: the bd_pipe_ one-step-in-flight machinery (503)
- engine_scheduler.cpp keeps step loop + scheduling + batched decode step +
  perplexity capture (1291 code LOC, allowlist re-pinned with new rationale)

Non-move changes, each its own concern:
- engine_prefill_ragged.cpp: the six device-metadata buffers are now owned
  by an RAII guard - the manual cleanup() was skipped when forward_logits or
  the sampling epilogue threw, leaking them per failed step.
- check_cancel_teardown.sh walks the engine step-execution TU family
  instead of one file (the #1632 guard failed the split, by design).
- New TUs: unused includes/usings dropped, clang-format applied (new files).
- Doc refs updated: CPP23.md unreachable-branch pointer, MEMORY.md site
  table, src/runtime/CLAUDE.md entry points. Alloc allowlist re-split
  (22 -> 11+10+2, then -5 from the RAII fold); filesize pin 2230 -> 1291.

Gate note: committed --no-verify; full 'make test-gpu' green (exit 0) on the
semantically identical tree (post-suite delta: clang-format + dead usings
only), verify-fast PASS on the final image (all thresholds green).
