+++ tools/alloc_allowlist.txt
@@ -5,7 +5,7 @@
 # be justified in review; the gate fails on any file that allocates and is
 # not listed, and equally on any listed file that no longer allocates.
 #
-# remaining: 66 files, 465 sites
+# remaining: 70 files, 478 sites
 
 src/compute/attention_cublas.cu  # 14
 src/compute/attention_fmha_sm120.cu  # 2
@@ -37,12 +37,12 @@ src/exec/executor_forward.cu  # 2
 src/exec/executor_forward_moe.cu  # 5
 src/exec/executor_forward_moe_batch.cu  # 7
 src/exec/executor_forward_moe_cutlass.cu  # 10
+src/exec/executor_gemm_dispatch.cu  # 4
 src/exec/executor_helpers.h  # 3
 src/exec/executor_lora.cu  # 2
 src/exec/executor_perplexity.cu  # 8
 src/exec/executor_ssm_gdn.cu  # 2
 src/exec/executor_workspace.cu  # 7
-src/exec/executor_gemm_dispatch.cu  # 4
 src/exec/executor_workspace_buffers.cu  # 17
 src/exec/expert_cache.cu  # 4
 src/exec/gemm_scratch.cu  # 1
@@ -66,12 +66,14 @@ src/quant/nvfp4_quant.cu  # 16
 src/runtime/batch.cpp  # 10
 src/runtime/cuda_graph.cu  # 26
 src/runtime/engine.cpp  # 6
+src/runtime/engine_decode_pipeline.cpp  # 2
 src/runtime/engine_graph_decode.cpp  # 30
 src/runtime/engine_internal.h  # 5
 src/runtime/engine_kv_cache_init.cpp  # 5
-src/runtime/engine_spec_mtp.cpp  # 2
-src/runtime/engine_prefill_ragged.cpp  # 12
-src/runtime/engine_scheduler.cpp  # 22
+src/runtime/engine_prefill.cpp  # 10
+src/runtime/engine_prefill_ragged.cpp  # 7
+src/runtime/engine_scheduler.cpp  # 10
 src/runtime/engine_spec_buffers.cpp  # 13
+src/runtime/engine_spec_mtp.cpp  # 2
 src/runtime/engine_workspace_warmup.cpp  # 2
 src/vision/vision_pipeline.cpp  # 2
